requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.36
libsql-client==0.3.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
gunicorn==21.2.0
